Move the task-key helpers into luxonis_ml.ldf.tasks
The task-key grammar belongs to the format, not to the data package. The
helpers now live beside the record model, and `luxonis_ml.data.utils` keeps
re-exporting them, so no import path breaks.

`get_task_group` moves here too. The augmentation engine held a private copy
of it, which kept every level of a nested task name. Only the engine could
use it, although the loader needs the same rule for sub-detections.
